// src/types/api.ts
export interface components {
  schemas: {
    BoardStatusBase: {
      /** Format: uuid */
      id: string;
      /** Format: uuid */
      boardId: string;
      name: string;
      description: string | null;
      position: number;
      /** Format: date-time */
      createdAt: string;
      /** Format: date-time */
      updatedAt: string;
    };
    Error: {
      error: string;
    };
    CreateBoardStatus: {
      name: string;
      description?: string;
      position?: number;
    };
    UpdateBoardStatus: {
      name?: string;
      description?: string | null;
      position?: number;
    };
    Board: {
      /** Format: uuid */
      id: string;
      name: string;
      description: string | null;
      /** Format: date-time */
      createdAt: string;
      /** Format: date-time */
      updatedAt: string;
    };
    BoardWithStatuses: components['schemas']['Board'] & {
      statuses: components['schemas']['BoardStatus'][];
    };
    BoardStatus: {
      /** Format: uuid */
      id: string;
      /** Format: uuid */
      boardId: string;
      name: string;
      description: string | null;
      position: number;
      /** Format: date-time */
      createdAt: string;
      /** Format: date-time */
      updatedAt: string;
      tasks: components['schemas']['Task'][];
    };
    Task: {
      /** Format: uuid */
      id: string;
      /** Format: uuid */
      boardStatusId: string;
      title: string;
      description: string | null;
      position: number;
      /** Format: date-time */
      createdAt: string;
      /** Format: date-time */
      updatedAt: string;
    };
    CreateBoard: {
      name: string;
      description?: string;
    };
    UpdateBoard: {
      name?: string;
      description?: string | null;
    };
    TaskWithSubTasks: components['schemas']['Task'] & {
      subTasks: components['schemas']['SubTask'][];
    };
    SubTask: {
      /** Format: uuid */
      id: string;
      /** Format: uuid */
      taskId: string;
      description: string;
      isCompleted: boolean;
      /** Format: date-time */
      createdAt: string;
      /** Format: date-time */
      updatedAt: string;
    };
    CreateTask: {
      /** Format: uuid */
      boardStatusId: string;
      title: string;
      description?: string;
      position: number;
    };
    UpdateTask: {
      /** Format: uuid */
      boardStatusId?: string;
      title?: string;
      description?: string | null;
      position?: number;
    };
    CreateSubTask: {
      /** Format: uuid */
      taskId: string;
      description: string;
      isCompleted?: boolean;
    };
    UpdateSubTask: {
      description?: string;
      isCompleted?: boolean;
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}